Move material type to inventoryStorage package
