+++ src/main/java/org/folio/innreach/client/InventoryClient.java
@@ -1,17 +1,15 @@
 package org.folio.innreach.client;
 
-import java.util.UUID;
-
+import org.folio.innreach.domain.dto.folio.inventory.InventoryInstanceDTO;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import org.folio.innreach.domain.dto.folio.inventory.InventoryInstanceDTO;
+import java.util.UUID;
 
 @FeignClient("inventory")
 public interface InventoryClient {
 
   @GetMapping("/inventory/instances/{id}")
   InventoryInstanceDTO getInstanceById(@PathVariable("id") UUID instanceId);
-
 }
