+++ src/main/java/org/folio/innreach/client/InventoryStorageClient.java
@@ -5,13 +5,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import org.folio.innreach.domain.dto.folio.ResultList;
-import org.folio.innreach.domain.dto.folio.inventory.MaterialTypeDTO;
+import org.folio.innreach.domain.dto.folio.inventoryStorage.MaterialTypeDTO;
 
 @FeignClient("inventory-storage")
 public interface InventoryStorageClient {
 
   @GetMapping("/material-types")
   ResultList<MaterialTypeDTO> getMaterialTypes(@RequestParam("query") String query, @RequestParam("limit") int limit);
-
 
 }
