+++ src/test/java/org/folio/innreach/fixture/ContributionFixture.java
@@ -14,7 +14,7 @@ import org.jeasy.random.EasyRandom;
 import org.jeasy.random.EasyRandomParameters;
 
 import org.folio.innreach.domain.dto.folio.ResultList;
-import org.folio.innreach.domain.dto.folio.inventory.MaterialTypeDTO;
+import org.folio.innreach.domain.dto.folio.inventoryStorage.MaterialTypeDTO;
 import org.folio.innreach.domain.entity.CentralServer;
 import org.folio.innreach.domain.entity.Contribution;
 import org.folio.innreach.external.dto.InnReachLocationDTO;
